refactor(cli): moved the CLI extraction out of the BugsnagCliTask and into a dedicated singleton
